requirements.txt
python-telegram-bot
supabase
openai
python-dotenv